Download some sources if $GET_FILES==true
If $GET_FILES is true in the environment, then use `spectool` (part of
`rpmdevtools`) to download all source files listed as URLs in the spec
file into the `upstream` directory.

This allows `build-from-github` to work when a spec file has additional
source files that are not part of the repo.

spectool skips downloading files that already exist so this should be
safe to run.